Reify names.Ident validity
